move plugin runner exceptions file to the plugin runner directory
